pyproject.toml
[tool.poetry]
name = "farm_registration"
version = "0.1.0"
description = ""
authors = ["HT Admin <admin@somecompany.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
ipython = "^7.26.0"
django-extensions = "^3.1.3"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
